refactor(file-picker): extract header tabs component
